Refactor: test를 위한 util 함수들 개별 파일로 분할 및 리팩터링
