+++ client/src/__tests__/test-utils/customRender.tsx
@@ -1,0 +1,20 @@
+import React, { FC, ReactElement } from 'react';
+import { render, RenderOptions } from '@testing-library/react';
+import { ThemeProvider } from 'styled-components';
+
+import GlobalStyle from '~styles/global-style';
+import { theme } from '~styles/theme';
+
+const AllTheProviders: FC<{ children: React.ReactNode }> = ({ children }) => {
+  return (
+    <ThemeProvider theme={theme}>
+      <GlobalStyle />
+      {children}
+    </ThemeProvider>
+  );
+};
+
+export const customRender = (
+  ui: ReactElement,
+  options?: Omit<RenderOptions, 'wrapper'>,
+) => render(ui, { wrapper: AllTheProviders, ...options });
