+++ client/src/__tests__/test-utils/index.tsx
@@ -1,0 +1,3 @@
+export * from '@testing-library/react';
+export { customRender as render } from './customRender';
+export { createMockRouter } from './createMockRouter';
